+++ examples/websocket/src/main.rs
@@ -55,7 +55,8 @@ fn main() {
             // websocket route
             .resource("/ws/", |r| r.method(Method::GET).f(ws_index))
             // static files
-            .handler("/", fs::StaticFiles::new("../static/", true)))
+            .handler("/", fs::StaticFiles::new("../static/", true)
+                     .index_file("index.html")))
         // start http server on 127.0.0.1:8080
         .bind("127.0.0.1:8080").unwrap()
         .start();